combine search specs
